Studio workspace layoutchanges (#1089)
* Change studio layout

Replace tabs with menus in studio. Move all logic to a single big container.
