+++ src/shared/styles/result-table.less
@@ -9,7 +9,12 @@
     justify-content: space-between;
 
     .search {
-      width: 50%;
+      width: 30%;
+    }
+
+    .title {
+      width: auto;
+      display: inline-block;
     }
 
     .select-column {
